+++ metadata.rb
@@ -1,21 +1,15 @@
 name 'poshbot'
 maintainer 'Garrett Yamada & Kyle Levenick'
 maintainer_email 'kyle@levenick.io'
-license 'All Rights Reserved'
+license 'MIT'
 description 'Installs/Configures PoshBot'
-long_description 'Installs/Configures PoshBot'
+long_description 'Installs and configures an instance of PoshBot, a PowerShell bot framework.'
 version '0.1.0'
+
+issues_url 'https://github.com/YourCrush/poshbot/issues'
+source_url 'https://github.com/YourCrush/poshbot'
+
+supports 'windows'
+
 chef_version '>= 13.0'
 depends 'nssm'
-
-# The `issues_url` points to the location where issues for this cookbook are
-# tracked.  A `View Issues` link will be displayed on this cookbook's page when
-# uploaded to a Supermarket.
-#
-# issues_url 'https://github.com/<insert_org_here>/its-poshbot/issues'
-
-# The `source_url` points to the development repository for this cookbook.  A
-# `View Source` link will be displayed on this cookbook's page when uploaded to
-# a Supermarket.
-#
-# source_url 'https://github.com/<insert_org_here>/its-poshbot'