+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-# its-poshbot CHANGELOG
+# poshbot CHANGELOG
 
-This file is used to list changes made in each version of the its-poshbot cookbook.
+This file is used to list changes made in each version of the poshbot cookbook.
 
-# 0.1.0
+# 1.0.0
 
 Initial release.
+
+# 1.1.0
+
+Mostly README changes.
+
+# 1.2.0 
+
+All configurable values are now attributes, even for the start-poshbot.ps1 template and any directory paths inside poshbot::poshbot.
+
+